dashboard/requirements.txt
dash
plotly
psycopg2-binary
pandas
dash-bootstrap-components
numpy
python-dotenv
ruff
sqlalchemy